requirements.txt
Django>=5.2.1
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-filter>=23.5
django-rest-framework>=0.1.0
django-storages>=1.14.2
django-templated-mail>=1.1.1
django-utils>=0.0.2
django-widget-tweaks>=1.5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
whitenoise>=6.6.0